Fix composer integration + Code fixes
Allow for <strong> and <em> tags with hack
