+++ composer.json
@@ -13,7 +13,7 @@
       }
   ],
   "require": {
-    "php": "^5.3.3 || ^7.0 || ^7.1 || ^7.2"
+    "php": "^7.0 || ^7.1 || ^7.2"
   },
   "autoload": {
       "psr-4": {
